add security, fix controllers with security

+++ src/main/java/com/eltavi/recipegram/controller/StepController.java
@@ -2,10 +2,15 @@ package com.eltavi.recipegram.controller;
 
 import com.eltavi.recipegram.dto.StepDto;
 import com.eltavi.recipegram.entity.FileTable;
+import com.eltavi.recipegram.entity.Recipe;
+import com.eltavi.recipegram.entity.Step;
 import com.eltavi.recipegram.exception.BadRequestException;
 import com.eltavi.recipegram.exception.NotFoundException;
+import com.eltavi.recipegram.exception.ResourceForbiddenException;
 import com.eltavi.recipegram.service.FileService;
+import com.eltavi.recipegram.service.RecipeService;
 import com.eltavi.recipegram.service.StepService;
+import com.eltavi.recipegram.service.UserService;
 import com.eltavi.recipegram.validator.StepDtoValidator;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.io.IOUtils;
@@ -20,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.security.Principal;
 import java.util.List;
 
 @RestController
@@ -29,6 +35,8 @@ public class StepController {
 
     private final StepService stepService;
     private final FileService fileService;
+    private final RecipeService recipeService;
+    private final UserService userService;
     private final StepDtoValidator stepDtoValidator;
 
     @GetMapping("steps")
@@ -47,8 +55,9 @@ public class StepController {
     }
 
     @PostMapping("steps")
-    public StepDto addStep(@RequestPart(name = "stepDto") StepDto stepDto,
-                           @RequestPart(name = "file") MultipartFile file) {
+    public StepDto addStep(Principal auth,
+                           @RequestPart(name = "stepDto") StepDto stepDto,
+                           @RequestPart(name = "file", required = false) MultipartFile file) {
         stepDtoValidator.validate(stepDto);
         if (file != null) {
             try {
@@ -60,10 +69,16 @@ public class StepController {
                 fileTable.setPhotoContentType(file.getContentType());
                 fileTable.setPhotoBlob(bytes);
 
+                Long userId = userService.findByUsername(auth.getName()).getId();
+                Recipe recipe = recipeService.findRecipeById(stepDto.getRecipeId());
+                if (!recipe.getUser().getId().equals(userId)) {
+                    throw new ResourceForbiddenException("You can't change other's recipes");
+                }
+
                 fileService.saveFile(fileTable);
                 stepDto.setImageStep(fileTable);
                 return stepService.save(stepDto);
-            } catch (Exception e) {
+            } catch (IOException e) {
                 throw new BadRequestException("You failed to upload => " + e.getMessage());
             }
         } else {
@@ -72,7 +87,8 @@ public class StepController {
     }
 
     @PatchMapping("steps/{id}")
-    public StepDto changeStep(@PathVariable Long id,
+    public StepDto changeStep(Principal auth,
+                              @PathVariable Long id,
                               @RequestPart(name = "stepDto") StepDto stepDto,
                               @RequestPart(name = "file", required = false) MultipartFile file) {
         stepDtoValidator.checkValidUpdateFields(stepDto);
@@ -86,6 +102,13 @@ public class StepController {
                 fileTable.setPhotoContentType(file.getContentType());
                 fileTable.setPhotoBlob(bytes);
 
+                Long userId = userService.findByUsername(auth.getName()).getId();
+                Step step = stepService.findStepById(id);
+                Recipe recipe = recipeService.findRecipeById(step.getRecipe().getId());
+                if (!recipe.getUser().getId().equals(userId)) {
+                    throw new ResourceForbiddenException("You can't change other's recipes");
+                }
+
                 Long fileId = stepService.deleteFile(id);
                 fileService.deleteFileById(fileId);
                 fileService.saveFile(fileTable);
@@ -97,12 +120,21 @@ public class StepController {
                 throw new BadRequestException("You failed to upload => " + e.getMessage());
             }
         } else {
+            stepDto.setId(id);
             return stepService.changeStep(stepDto);
         }
     }
 
     @DeleteMapping("{id}")
-    public void deleteStep(@PathVariable Long id) {
+    public void deleteStep(Principal auth,
+                           @PathVariable Long id) {
+        Long userId = userService.findByUsername(auth.getName()).getId();
+        Step step = stepService.findStepById(id);
+        Recipe recipe = recipeService.findRecipeById(step.getRecipe().getId());
+        if (!recipe.getUser().getId().equals(userId)) {
+            throw new ResourceForbiddenException("You can't change other's recipes");
+        }
+
         if (!stepService.deleteStep(id)) {
             throw new NotFoundException("There is no step with this id");
         } else {
